feat(workflow): move dagre layout helpers into lib graph-edit
