+++ packages/lib/package.json
@@ -1396,6 +1396,18 @@
       "import": "./dist/workflows/index.mjs",
       "default": "./src/workflows/index.ts"
     },
+    "./workflows/graph-edit/layout": {
+      "types": "./src/workflows/graph-edit/layout.ts",
+      "source": "./src/workflows/graph-edit/layout.ts",
+      "import": "./dist/workflows/graph-edit/layout.mjs",
+      "default": "./src/workflows/graph-edit/layout.ts"
+    },
+    "./workflows/graph-edit/layout-constants": {
+      "types": "./src/workflows/graph-edit/layout-constants.ts",
+      "source": "./src/workflows/graph-edit/layout-constants.ts",
+      "import": "./dist/workflows/graph-edit/layout-constants.mjs",
+      "default": "./src/workflows/graph-edit/layout-constants.ts"
+    },
     "./workflows/variable-ref-rewriter": {
       "types": "./src/workflows/variable-ref-rewriter.ts",
       "source": "./src/workflows/variable-ref-rewriter.ts",
@@ -1453,6 +1465,7 @@
     "commander": "^12.0.0",
     "cronstrue": "^3.14.0",
     "csv-parser": "^3.2.0",
+    "dagre": "^0.8.5",
     "date-fns": "catalog:",
     "date-fns-tz": "^3.2.0",
     "decode-ico": "^0.4.0",
@@ -1516,6 +1529,7 @@
     "@auxx/typescript-config": "workspace:*",
     "@faker-js/faker": "catalog:",
     "@types/addressparser": "^1.0.3",
+    "@types/dagre": "^0.7.52",
     "@types/email-reply-parser": "^1.4.2",
     "@types/html-to-text": "^9.0.4",
     "@types/jsdom": "^21.1.7",
